WIP: ec2_vol.py - Check if instance is found, fail if not
Add check to see if we get nothing back from querying the instance and fail with a descriptive error message if we dont get what we expect.

I ran into this issue when testing, I had entered the wrong region and it was searching for the instance ID in the wrong region. Since it never checks if we got valid data back, when we try and get an element from the NoneType, we get an error "TypeError: 'NoneType' object is not subscriptable." It took a very long time to diagnose this and would be easier if I had a more descriptive error message.